The ID_refl_... is not necessary in any case
The macro INIZ_ClassJc is reduced. Some check macros are added,
for example INSTANCEOF_ObjectJc(OTHIZ, REFL).
It is possible to dismiss using ClassJc reflection instances for simple systems with less code.
The type check is still possible using the ID_refl_.... with that macros.
But if DEF_ObjectJc_RERFLREF is set, the ID_refl_... is not necessary.
This is important because large software systems with some more types
do not need effort for some not used ID_refl_...