bugfix memory destroy on ctor_ObjectJc(), tested and found in Linux gcc
* The meaning of the size information was obscure, for C usage it was the size of the
C plain data and similar the size of the instance.
Hence a memset(othis, 0, size) was part of the software for some years.
* But it should be used as information about the instance's size also of C++ class instances.
Hence the memset was faulty. In generally memset should not be used as a watering can.
It should never be uses, instead a dedicated set to 0 of all variables may be more obviously
also for the documentation of the program.

* some fixes while testing RingBuffer

* some fixes while using gcc in Linux, not all compiles yet. Especially static inline problematic.

* Using emC_inclComplSpec/cc_Gcc64 for gcc on 64-bit-Linux